# perf(python): Much faster `Series` construction from subclasses of standard Python types