@ritchie46 2024-12-05
I see you learned some Dutch.